octeontx2-af: npa: cn20k: Fix DPC config mbox
The DPC configuration that is written into the NPA_AF_DPCX_CFG
register should be 64 bit in size instead of 16 bit. Fix the
DPC alloc mbox accordingly.

Fixes: 600e2df1c8ab ("octeontx2-af: npa: cn20k: Fix DPC config mbox")

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/168635
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
(cherry picked from commit 6c856800a2bc06d673c1f1cdd182321a82f38ade)
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/168946

Orabug: 39318904

+++ drivers/net/ethernet/marvell/octeontx2/af/mbox.h
@@ -1033,7 +1033,7 @@ struct npa_cn20k_aq_enq_rsp {
 
 struct npa_cn20k_dpc_alloc_req {
 	struct mbox_msghdr hdr;
-	u16 dpc_conf;
+	u64 dpc_conf;
 };
 
 struct npa_cn20k_dpc_alloc_rsp {
